+++ packages/amplify-graphql-api-construct-tests/src/__tests__/backends/relationships/app.ts
@@ -2,7 +2,7 @@
 import 'source-map-support/register';
 import { App, Stack, Duration } from 'aws-cdk-lib';
 // @ts-ignore
-import { AmplifyGraphqlApi, AmplifyGraphqlSchema } from '@aws-amplify/graphql-construct-alpha';
+import { AmplifyGraphqlApi, AmplifyGraphqlDefinition } from '@aws-amplify/graphql-api-construct';
 
 const packageJson = require('../package.json');
 
@@ -12,7 +12,7 @@ const stack = new Stack(app, packageJson.name.replace(/_/g, '-'), {
 });
 
 new AmplifyGraphqlApi(stack, 'GraphqlApi', {
-  schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+  definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
     type Blog @model @auth(rules: [{ allow: public }]) {
       title: String!
       description: String!
@@ -39,7 +39,7 @@ new AmplifyGraphqlApi(stack, 'GraphqlApi', {
       posts: [Post] @manyToMany(relationName: "PostTags")
     }
   `),
-  authorizationConfig: {
+  authorizationModes: {
     apiKeyConfig: { expires: Duration.days(7) },
   },
 });